made saved foods a component to be used
